Add Omniauth support for SAML and Google OAuth2 with MultiAuth
This commit adds the ability for users to authenticate via SAML or Google OAuth2 in addition to the existing authentication method. This includes modifications to the User model to accept omniauth providers, creation of OmniauthCallbacks and Sessions controllers for users, and updates to views and routes to handle these new authentication methods. Additionally, required Omniauth gems have been added to the Gemfile.